STAC-21755 Change settings backup to use new mechanism

+++ stable/suse-observability/scripts/restore-configuration-backup.sh
@@ -1,18 +1,27 @@
 #!/usr/bin/env bash
 set -Eeuo pipefail
 
-TMP_DIR=/tmp
+export BACKUP_DIR=/settings-backup-data
+export TMP_DIR=/tmp
 
-export AWS_ACCESS_KEY_ID
-AWS_ACCESS_KEY_ID="$(cat /aws-keys/accesskey)"
-export AWS_SECRET_ACCESS_KEY
-AWS_SECRET_ACCESS_KEY="$(cat /aws-keys/secretkey)"
+RESTORE_FILE="${BACKUP_DIR}/${BACKUP_FILE}"
 
-echo "=== Downloading Configuration backup \"${BACKUP_FILE}\" from bucket \"${BACKUP_CONFIGURATION_BUCKET_NAME}\"..."
-sts-toolbox aws s3 --endpoint "http://${MINIO_ENDPOINT}" --region minio cp "s3://${BACKUP_CONFIGURATION_BUCKET_NAME}/${BACKUP_CONFIGURATION_S3_PREFIX}${BACKUP_FILE}" "${TMP_DIR}/${BACKUP_FILE}"
+if [ "$BACKUP_CONFIGURATION_UPLOAD_REMOTE" == "true" ] && [ ! -f "${RESTORE_FILE}" ]; then
+  export AWS_ACCESS_KEY_ID
+  AWS_ACCESS_KEY_ID="$(cat /aws-keys/accesskey)"
+  export AWS_SECRET_ACCESS_KEY
+  AWS_SECRET_ACCESS_KEY="$(cat /aws-keys/secretkey)"
 
-echo "=== Deleting preinstalled stackpacks..."
-sts stackpack list --url "${STACKSTATE_ROUTER_ENDPOINT}" --installed --output json | jq '.stackpacks[] | (.name + " " + (.configurations[0].id | tostring) )' -r | while read -r NAME ID ; do sts stackpack uninstall -n "${NAME}" -i "${ID}" --url "${STACKSTATE_ROUTER_ENDPOINT}" ; done
-echo "=== Importing Configuration data from \"${BACKUP_FILE}\"..."
-sts settings apply --namespace "" --url "${STACKSTATE_ROUTER_ENDPOINT}" --file "${TMP_DIR}/${BACKUP_FILE}"
+  echo "=== Downloading Settings backup \"${BACKUP_FILE}\" from bucket \"${BACKUP_CONFIGURATION_BUCKET_NAME}\"..."
+  sts-toolbox aws s3 --endpoint "http://${MINIO_ENDPOINT}" --region minio cp "s3://${BACKUP_CONFIGURATION_BUCKET_NAME}/${BACKUP_CONFIGURATION_S3_PREFIX}${BACKUP_FILE}" "${TMP_DIR}/${BACKUP_FILE}"
+  RESTORE_FILE="${TMP_DIR}/${BACKUP_FILE}"
+fi
+
+ if [ ! -f "${RESTORE_FILE}" ]; then
+  echo "=== Backup file \"${RESTORE_FILE}\" not found, exiting..."
+  exit 1
+ fi
+
+echo "=== Restoring settings backup from \"${BACKUP_FILE}\"..."
+/opt/docker/bin/settings-backup -Dlogback.configurationFile=/opt/docker/etc_log/logback.groovy -restore "${RESTORE_FILE}"
 echo "==="
